requirements.txt
langchain==0.0.141
Pillow==9.5.0
PyPDF2==3.0.1
streamlit==1.10.0
openai==0.27.4
tiktoken==0.3.3
faiss-cpu==1.7.3